Extract some hyper-independent header processing to the lowlevel crate
