+++ websocket-lowlevel/Cargo.toml
@@ -18,6 +18,8 @@ license = "MIT"
 byteorder = "1.0"
 rand = "0.6.1"
 bitflags = "1.0.4"
+base64 = "0.10.0"
+sha1 = "0.6"
 bytes = { version = "0.4", optional = true }
 futures = { version = "0.1", optional = true }
 native-tls = { version = "0.2.1", optional = true }
