+++ websocket-lowlevel/src/lib.rs
@@ -30,6 +30,7 @@ pub mod message;
 pub mod result;
 pub mod ws;
 pub mod stream;
+pub mod header;
 
 pub use message::OwnedMessage;
 pub use message::Message;
